Middleware that add an Etag header to the response

This middleware handles the If-None-Match headers, and it sets the status code of the response to 304 if If-None-Match matches the Etag. It also removes the response body in this case.

Usage

mw_etag(algorithm = "crc32")

Arguments

algorithm

Checksum algorithm to use. Only "crc32" is implemented currently.

Value

Handler function.

See also

Examples

app <- new_app()
app$use(mw_etag())
app
#> <webfakes_app>
#> routes:
#>   use *
#> fields and methods:
#>   all(path, ...)         # add route for *all* HTTP methods
#>   delete(path, ...)      # add route for DELETE
#>   engine(ext, engine)    # add template engine for file extension
#>   head(path, ...)        # add route for HEAD
#>   listen(port)           # start web app on port
#>   patch(path, ...)       # add route for PATCH
#>   post(path, ...)        # add route for POST
#>   put(path, ...)         # add route for PUT
#>   use(...)               # add middleware
#>   locals                 # app-wide shared data
#> # see ?webfakes_app for all methods