Parse a multipart HTTP request body

Adds the parsed form fields in the form element of the request and the parsed files to the files element.

mw_multipart(type = "multipart/form-data")

Arguments

type
Content type to match before parsing. If it does not match, then the request object is not modified.

Value

Handler function.

See also

Examples

app <- new_app()
app$use(mw_multipart())
app

#> <webfakes_app>
#> routes:
#>   use *
#> fields and methods:
#>   all(path, ...)         # add route for *all* HTTP methods
#>   delete(path, ...)      # add route for DELETE
#>   engine(ext, engine)    # add template engine for file extension
#>   head(path, ...)        # add route for HEAD
#>   listen(port)           # start web app on port
#>   patch(path, ...)       # add route for PATCH
#>   post(path, ...)        # add route for POST
#>   put(path, ...)         # add route for PUT
#>   use(...)               # add middleware
#>   locals                 # app-wide shared data
#> # see ?webfakes_app for all methods